Move skill layout tests out of synchronizer fixture
Separate layout and catalog behavior tests from ToolSkillSynchronizerTests so the synchronizer fixture now focuses on synchronization responsibilities after the class split.
