+++ Assets/Tests/Editor/SkillInstallLayoutTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0dc989fa0ebc44cdb791df127f1584e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
